# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cosmic Webapp",
    "version": "1.0.0",
    "description": "API documentation with instructions"
  },
  "servers": [
    {
      "url": "/cosmic/"
    },
    {
      "url": "/"
    }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Register a new user",
        "description": "Registers a new user",
        "tags": ["Basic Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "first_name": { "type": "string", "example": "user" },
                  "last_name": { "type": "string", "example": "user" },
                  "email": { "type": "string", "example": "mike_w@gmail.com" },
                  "clerk_id": { "type": "string", "example": "clerk_id_12345" },
                  "role": { "type": "string", "example": "user" },
                  "username": { "type": "string", "example": "username" }
                },
                "required": ["email", "clerk_id"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/auth/user": {
      "get": {
        "summary": "Get authenticated user info",
        "tags": ["Basic Auth"],
        "security": [
          {
            "ACCESS_TOKEN": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated user details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "email": { "type": "string", "format": "email" },
                        "role": { "type": "string" },
                        "created_at": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updated_at": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing access token"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/auth/users": {
      "get": {
        "summary": "Get list of all users",
        "tags": ["Basic Auth"],
        "security": [
          {
            "ACCESS_TOKEN": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error retrieving users"
          }
        }
      }
    },
...
    "/activity": {
      "get": {
        "summary": "Get all activites",
        "tags": ["Users"],
        "security": [
          {
            "ACCESS_TOKEN": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of activites for task"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ACCESS_TOKEN": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "REFRESH_TOKEN": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}